# Use inline cache for super() calls