This introduces inline caching for `super()` calls. This is able to skip the method lookup on the superclass but is still slower than a standard method call as some checks (ie. `rb_obj_is_kind_of`) are still required.

## Benchmarks

```
require "benchmark/ips"

class A;     def foo; end; def bar; end; end
class B < A; def foo; super; end; end

Benchmark.ips do |x|
  obj = B.new
  x.report("foo") { obj.foo }
  x.report("bar") { obj.bar }
end
```

**Before**

```
                 foo     12.558M (± 1.6%) i/s -     63.187M in   5.033012s
                 bar     25.596M (± 1.0%) i/s -    128.195M in   5.008823s
```

**After**

```
                 foo     17.329M (± 0.5%) i/s -     86.647M in   5.000196s
                 bar     25.564M (± 0.4%) i/s -    128.370M in   5.021504s
```

## Rails benchmark

This makes `ActionDispatch::Request.new` faster, as it calls `super()` a few times.

```
$ cat benchmark_build_request.rb
# frozen_string_literal: true

require "action_dispatch"
require "benchmark/ips"

REQUEST_ENV = Rack::MockRequest.env_for("http://example.org/hello_world/json")
REQUEST_ENV.freeze

Benchmark.ips do |x|
  x.report "ActionDispatch::Request.new", <<~RUBY
    ActionDispatch::Request.new(REQUEST_ENV)
  RUBY
end
```

**Before**
```
ActionDispatch::Request.new
                          2.500M (± 2.2%) i/s -     12.609M in   5.045582s

```

**After**
```
ActionDispatch::Request.new
                          3.371M (± 0.4%) i/s -     16.898M in   5.012308s
```